backend/django_app/requirements.txt
django==3.1
djangorestframework==3.11.1
django-rest-auth==0.9.5
django-cors-headers==3.5.0
psycopg2-binary==2.8.5
pandas==1.1.1
Pillow==8.2.0
joblib==0.16.0
gunicorn==20.0.4
Keras==2.4.3
tqdm==4.61.0
tensorflow-cpu==2.5.0
numpy==1.19.5
scikit-learn==0.23.2
scipy==1.5.4